refactor: naming pattern, re-org hooks folder for clarity
