+++ apps/ensadmin/src/app/name/[name]/page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { Card, CardContent } from "@/components/ui/card";
-import { useActiveNamespace } from "@/hooks/use-active-namespace";
 import { useRecords } from "@ensnode/ensnode-react";
 import { ResolverRecordsSelection, getCommonCoinTypes } from "@ensnode/ensnode-sdk";
 import { useParams } from "next/navigation";
 
+import { useActiveNamespace } from "@/hooks/active/use-active-namespace";
 import { AdditionalRecords } from "./_components/AdditionalRecords";
 import { Addresses } from "./_components/Addresses";
 import { NameDetailPageSkeleton } from "./_components/NameDetailPageSkeleton";
