+++ apps/ensadmin/src/components/providers/active-ensnode-provider.tsx
@@ -1,10 +1,11 @@
 "use client";
 
-import { useActiveENSNodeUrl } from "@/hooks/use-active-ensnode-url";
+import { useActiveENSNodeUrl } from "@/hooks/active/use-active-ensnode-url";
 import { ENSNodeProvider as _ENSNodeProvider } from "@ensnode/ensnode-react";
 import { PropsWithChildren } from "react";
 
-export function ENSNodeProvider({ children }: PropsWithChildren) {
+export function ActiveENSNodeProvider({ children }: PropsWithChildren) {
   const url = useActiveENSNodeUrl();
+
   return <_ENSNodeProvider config={{ client: { url } }}>{children}</_ENSNodeProvider>;
 }
