+++ apps/ensadmin/src/hooks/active/use-active-namespace.ts
@@ -1,4 +1,4 @@
-import { useActiveENSNodeConfig } from "@/hooks/use-active-ensnode-config";
+import { useActiveENSNodeConfig } from "./use-active-ensnode-config";
 
 // helper for accessing the active ENSNode's Config's Namespace
 export const useActiveNamespace = () => useActiveENSNodeConfig().namespace;
