+++ apps/ensadmin/src/hooks/async/use-ens-app-name-url.ts
@@ -1,0 +1,12 @@
+import { getExternalEnsAppNameUrl } from "@/lib/namespace-utils";
+import { Name } from "@ensnode/ensnode-sdk";
+import { useNamespace } from "./use-namespace";
+
+export function useENSAppNameUrl(name: Name) {
+  const { data: namespace, ...query } = useNamespace();
+
+  return {
+    ...query,
+    data: namespace ? getExternalEnsAppNameUrl(name, namespace) : null,
+  };
+}
