--- apps/ensadmin/src/hooks/use-ens-app-name-url.ts
@@ -1,12 +1,0 @@
-import { getExternalEnsAppNameUrl } from "@/lib/namespace-utils";
-import { useENSIndexerConfig } from "@ensnode/ensnode-react";
-import { Name } from "@ensnode/ensnode-sdk";
-
-export function useENSAppNameUrl(name: Name) {
-  const { data, ...query } = useENSIndexerConfig();
-
-  return {
-    ...query,
-    data: data?.namespace ? getExternalEnsAppNameUrl(name, data.namespace) : null,
-  };
-}
